[AOT Refactor] runtime wrappers
[ghstack-poisoned]
